Achievement 2/Exercise2.7/src/requirements/dev.txt
-r base.txt
ipython==8.18.1
django-extensions==3.2.3
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==6.1.0
colorama==0.4.6
decorator==5.1.1
django-markdownify==0.9.5
django-markdownx==4.0.7
executing==2.0.1
importlib_metadata==8.4.0
jedi==0.19.1
Markdown==3.7
matplotlib-inline==0.1.7
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.2.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.20.1
